Update the rbenv-installation-dir to pull from RBENV_ROOT.
The RBENV_ROOT environment variable is actually set by rbenv. For users
of GitHub's Boxen project this is installed to
/opt/boxen/rbenv. Additionally the same problem becomes present on
corporate machines where often we only want to install rubies once.

During code review @senny made a good point: it may not be in the
environment and we would still want to load it up in Emacs. This
by his suggestion seems like the right patch to me.